pyproject.toml
[tool.poetry]
name = "spotify-ai"
version = "0.1.0"
description = ""
authors = ["Joy Cheng <joy.cheng@phntms.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
spotipy = "^2.24.0"
langchain = "^0.2.11"
langchain-community = "^0.2.10"
boto3 = "^1.34.149"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
